competitor_research/requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
requests>=2.31.0
openpyxl>=3.1.0
python-docx>=1.0.0
